refactor(otlp): separate retry policy from engine

Copilot-Session: 0bc45e17-c983-47d7-9a97-f6b1909ceefb
